feat: Add validation tests for MQTT v5 protocol features
- Implement tests for protocol name and level validation.
- Add tests for reason code validation for various packet types (ConnAck, SubAck, Auth).
- Include tests for fixed header flags validation.
- Validate property scopes for different packet types (Publish, Connect, Auth).
- Test credential validation for ConnectPacket with password without username.
- Implement shared subscription parsing tests, covering valid and invalid cases.
- Add tests for publish packet validation and QoS packet ID mismatch.
- Validate shared subscription topic filter rules.

+++ mqtt_codec/src/v4/decoder.rs
@@ -61,7 +61,7 @@ impl Decoder for MqttDecoder {
 
         // Parse fixed header
         let first_byte = src[0];
-        let packet_type = (first_byte >> 4) as u8;
+        let packet_type = first_byte >> 4;
         let flags = first_byte & 0x0F;
 
         // Validate packet type (1-14, but not 0 or 15)
